requirements.txt
transformers
datasets
evaluate
accelerate
torch
numpy
python-dotenv